pyproject.toml
[tool.poetry]
name = "token-auth-proj"
version = "0.1.0"
description = "Test task: JWT Auth with Redis, role-based access"
authors = ["ludaNOFX <kozhaniichelovek322@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
redis = "^5.0.3"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pydantic-settings = "^2.8.1"
sqlalchemy = "^2.0.39"
alembic = "^1.15.1"
asyncpg = "^0.30.0"
databases = "^0.9.0"
pyclean = "^3.1.0"
argon2-cffi = "^23.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
mypy = "^1.8.0"
black = "^24.2.0"
isort = "^5.13.2"
ruff = "^0.3.2"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
  "ignore::DeprecationWarning",
]

[tool.ruff]
ignore = []
select = ["E", "F", "B", "UP", "SIM"]
line-length = 90
target-version = "py310"
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
extend-exclude = ["tests"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "E501"]
"deps.py" = ["F401"]
"tests/updated/conftest.py" = ["F403"]

[tool.ruff.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query", "fastapi.Header"]

[tool.ruff.mccabe]
max-complexity = 10

[tool.black]
line-length = 90
target-version = ['py310']
